semi-automated-update

+++ shutdown/myshutdown.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 
-# shutdown script for lxde
+# shutdown script (it has to be manually invoked)
 # add symlink to this file in ~/.local/bin which should be in your path
 # ln -s $MYDATA/git/bin-git/shutdown/myshutdown.sh myshutdown.sh
 
@@ -14,8 +14,15 @@ gitsyncall;
 
 #BOUND END   : addCommandsToRun
 
-############## power off !!!!!!!!!
-echo "$(date) myshutdown.sh end BYE!!" >> $MYDATA/local/logs/mysystem.log
-poweroff
+############## take final permission to shutdown
+echo -n "Shall I shutdown? (y/n) ";
+read response;
+if [[ $response == "y" || $response == "Y" ]]; then
+  ############## power off !!!!!!!!!
+  echo "$(date) myshutdown.sh end BYE!!" >> $MYDATA/local/logs/mysystem.log
+  poweroff
+fi
+
+
